consolidate SentryThreadWrapper.onMainThread to SentryDispatchQueueWrapper.dispatchSyncOnMainQueue
