+++ SentryTestUtils/TestSentryDispatchQueueWrapper.swift
@@ -34,13 +34,6 @@ public class TestSentryDispatchQueueWrapper: SentryDispatchQueueWrapper {
     public var blockOnMainInvocations = Invocations<() -> Void>()
     public var blockBeforeMainBlock: () -> Bool = { true }
 
-    public override func dispatchOnMainQueue(block: @escaping () -> Void) {
-        blockOnMainInvocations.record(block)
-        if blockBeforeMainBlock() {
-            block()
-        }
-    }
-
     public override func dispatchAsyncOnMainQueue(block: @escaping () -> Void) {
         blockOnMainInvocations.record(block)
         if blockBeforeMainBlock() {
