+++ Sources/Sentry/Profiling/SentryProfilerState.mm
@@ -1,10 +1,11 @@
 #import "SentryProfilerState.h"
 #if SENTRY_TARGET_PROFILING_SUPPORTED
 #    import "SentryBacktrace.hpp"
+#    import "SentryDependencyContainer.h"
+#    import "SentryDispatchQueueWrapper.h"
 #    import "SentryFormatter.h"
 #    import "SentryProfileTimeseries.h"
 #    import "SentrySample.h"
-#    import "SentryThreadWrapper.h"
 #    import <mach/mach_types.h>
 #    import <mach/port.h>
 #    import <mutex>
@@ -64,7 +65,8 @@ parseBacktraceSymbolsFunctionName(const char *symbol)
     if (self = [super init]) {
         _mutableState = [[SentryProfilerMutableState alloc] init];
         _mainThreadID = 0;
-        [SentryThreadWrapper onMainThread:^{ [self cacheMainThreadID]; }];
+        [SentryDependencyContainer.sharedInstance.dispatchQueueWrapper
+            dispatchSyncOnMainQueue:^{ [self cacheMainThreadID]; }];
     }
     return self;
 }
