+++ Sources/Sentry/SentryDispatchQueueWrapper.m
@@ -1,4 +1,5 @@
 #import "SentryDispatchQueueWrapper.h"
+#import "SentryDependencyContainer.h"
 #import "SentryThreadWrapper.h"
 #import <Foundation/Foundation.h>
 
@@ -41,11 +42,6 @@ NS_ASSUME_NONNULL_BEGIN
             block();
         }
     });
-}
-
-- (void)dispatchOnMainQueue:(void (^)(void))block
-{
-    [SentryThreadWrapper onMainThread:block];
 }
 
 - (void)dispatchSync:(void (^)(void))block
