+++ Sources/Sentry/SentryThreadWrapper.m
@@ -20,17 +20,6 @@ NS_ASSUME_NONNULL_BEGIN
     // No op. Only needed for testing.
 }
 
-+ (void)onMainThread:(void (^)(void))block
-{
-    if ([NSThread isMainThread]) {
-        SENTRY_LOG_DEBUG(@"Already on main thread.");
-        block();
-    } else {
-        SENTRY_LOG_DEBUG(@"Dispatching asynchronously to main queue.");
-        dispatch_async(dispatch_get_main_queue(), block);
-    }
-}
-
 @end
 
 NS_ASSUME_NONNULL_END
