+++ Sources/Sentry/include/SentryDispatchQueueWrapper.h
@@ -19,8 +19,6 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)dispatchAsyncOnMainQueue:(void (^)(void))block
     NS_SWIFT_NAME(dispatchAsyncOnMainQueue(block:));
 
-- (void)dispatchOnMainQueue:(void (^)(void))block NS_SWIFT_NAME(dispatchOnMainQueue(block:));
-
 - (void)dispatchSyncOnMainQueue:(void (^)(void))block
     NS_SWIFT_NAME(dispatchSyncOnMainQueue(block:));
