+++ Tests/SentryTests/Networking/SentryDispatchQueueWrapperTests.swift
@@ -26,5 +26,29 @@ class SentryDispatchQueueWrapperTests: XCTestCase {
         XCTAssertTrue(firstWasCalled)
         XCTAssertFalse(secondWasCalled)
         XCTAssertTrue(thirdWasCalled)
+    }    
+    
+    func testDispatchSyncToMainThreadFromNonMainContext() {
+        let e = expectation(description: "Asserted that execution happened on main thread")
+        let sut = SentryDispatchQueueWrapper()
+        sut.dispatchSyncOnMainQueue {
+            XCTAssertTrue(Thread.isMainThread)
+            e.fulfill()
+        }
+        waitForExpectations(timeout: 1)
+    }
+
+    func testDispatchSyncToMainQueueFromNonMainContext() {
+        let e = expectation(description: "Asserted that execution happened on main thread")
+        let q = DispatchQueue(label: "a nonmain queue", qos: .background)
+        q.async {
+            XCTAssertFalse(Thread.isMainThread)
+            let sut = SentryDispatchQueueWrapper()
+            sut.dispatchSyncOnMainQueue {
+                XCTAssertTrue(Thread.isMainThread)
+                e.fulfill()
+            }
+        }
+        waitForExpectations(timeout: 1)
     }
 }
